+++ lua/obsidian/lsp/code_action.lua
@@ -1,0 +1,5 @@
+-- ObsidianExtractNote
+-- ObsidianQuickSwitch
+-- ObsidianNew
+-- ObsidianBacklinks
+-- ObsidianDailies