+++ lua/obsidian/lsp/handlers/resolve.lua
@@ -1,0 +1,9 @@
+return function(_, params, handler, _)
+  params.documentation = {
+    value = [[# Heading 1
+[link](https://example.com)
+     ]],
+    kind = "markdown",
+  }
+  handler(nil, params)
+end