+++ lua/obsidian/lsp/handlers/workplace_diagnostics.lua
@@ -1,0 +1,4 @@
+return function(_, params, handler, _)
+  print "diagnosticing workspace"
+  return
+end