+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `obsidian.config` type for user config type check.
 - More informative healthcheck.
 - A guide to embed images for both viewing in neovim and obsidian app: https://github.com/obsidian-nvim/obsidian.nvim/wiki/Images
+- Added a builtin lsp to handle completion, hover, diagnostics and code actions
 
 ### Changed